fix(FormRenameBranch): Run git interactively
